requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
llama-index>=0.10.0
llama-index-readers-file>=0.1.0
python-dotenv>=1.0.0
openai>=1.10.0